Update configuration files and packages
This commit primarily updates various configuration details and switches around some package preferences. Notably, in the home.nix file, a condition was set to include a specific path when in the ~/src/teleclinic/ git directory, and the syntax for enabling autosuggestions in zsh was corrected. Furthermore, some changes were made to the tracked files, with some shell scripts now tracked and the addition of .idea/ files to the .gitignore. Various dependencies and packages were updated in flake.lock and packages.nix.